Divide workflow into consumer/scheduler
The way things worked before, if I wanted to track replies for multiple
tweets, I would need to set up two lambda functions (one for regular
replies, one for quoted replies) per tracked tweet.  This worked ok
for the first one or two, but it's clearly a nuisance.

So now, as of this commit, the way things work is that there are two
jobs: a scheduler and a consumer.  The scheduler adds data to a Kinesis
stream, and the consumer is triggered by that data and uses the payload
to figure out which tweets to track.  This way, if I want to add a new
tweet to track, I need only make a small alteration to the scheduler.

In the future, the scheduler may (let's be honest - will probably) do
things like read the list of target tweets out of a database to avoid
needing a deploy to track a new tweet, as well as perform some *actual*
scheduling to make sure I don't go over the API rate limit.  The
workflow may also change to do things like only requesting a single page
of search results per consumer call.