Pad status IDs in DynamoDB to 32 zeroes
I have an upcoming change planned that requires the code to be able to
quickly find the highest status ID for a partition - currently, we're
storing status IDs as strings in DynamoDB, which means that if I ask
for the highest status ID, it's going to give me the highest by
lexicographical sort.  I thought about making a new table using a
numeric value for status IDs, but DynamoDB numeric values are floating
point and not precise enough.

So the way I plan on getting around this is padding Twitter status IDs
with zeroes - that will make the table sort the way I want with just a
small amount of data processing